+++ src/main/java/cl/m5d12/dao/Capacitacion.java
@@ -49,7 +49,7 @@ public class Capacitacion  {
 	
 	//Getters and Setters
 
-	public long getidCapa() {
+	public int getidCapa() {
 		return idCapa;
 	}
